carrot/requirements.txt
# Main project dependencies
# Add these for Ghibli AI image generation
torch==2.3.1
diffusers==0.29.2
accelerate==0.33.0
safetensors==0.4.4
transformers==4.43.3
Pillow==10.4.0
numpy==1.26.4
